repo-monitor/requirements.txt
requests>=2.28.0
PyYAML>=6.0
sendgrid>=6.11.0
python-dotenv>=1.0.0
certifi>=2024.0.0
